+++ NEWS.md
@@ -8,6 +8,7 @@
   - Add equality (`==`) and approximate equality (`isapprox`) methods for `NWState` and `NWParameter`
   - Minor documentation fixes and spelling corrections
   - fix [#310](https://github.com/JuliaDynamics/NetworkDynamics.jl/pull/310) (allow `VPIndex(i)` to index into network objects)
+  - implement [#307](https://github.com/JuliaDynamics/NetworkDynamics.jl/pull/307), allow nw[VIndex(:)], nw[[VIndex(1),VIndex(2)]]
 
 ## v0.10.4 Changelog
 - [#303](https://github.com/JuliaDynamics/NetworkDynamics.jl/pull/303) update for ModelingToolkit.jl v10 compatibility: